[HSC-156] 지역별 데이터 평균량 조회 API -> 데이터 평균량 소수점으로 수정

+++ src/main/java/site/holliverse/admin/web/dto/analytics/AdminRegionalMetricResponseDto.java
@@ -27,7 +27,7 @@ public record AdminRegionalMetricResponseDto(
             String regionCode,
             String region,
             long averageSales,
-            long averageDataUsageGb
+            double averageDataUsageGb
     ) {
     }
